maple-loops/HelloWorld/requirements.txt
# used robosapiensIO version
robosapiensio==0.3.9
# application-specific
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.2
packaging==24.1
paho-mqtt==2.1.0
pillow==11.0.0
portion==2.6.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
six==1.16.0
sortedcontainers==2.4.0



